+++ picoserve/src/response/with_state.rs
@@ -70,3 +70,55 @@ impl<T: IntoResponse, State> IntoResponseWithState<State> for T {
         self.write_to(connection, response_writer).await
     }
 }
+
+/// A response which also updates the state. Returned by [`WithStateUpdate::with_state_update`]
+pub struct IntoResponseWithStateUpdate<
+    State,
+    T: IntoResponseWithState<State>,
+    F: AsyncFnOnce(&State),
+> {
+    response: T,
+    state_update: F,
+    _state: core::marker::PhantomData<fn(&State)>,
+}
+
+impl<State, T: IntoResponseWithState<State>, F: AsyncFnOnce(&State)> IntoResponseWithState<State>
+    for IntoResponseWithStateUpdate<State, T, F>
+{
+    async fn write_to_with_state<R: Read, W: ResponseWriter<Error = R::Error>>(
+        self,
+        state: &State,
+        connection: Connection<'_, R>,
+        response_writer: W,
+    ) -> Result<ResponseSent, W::Error> {
+        let Self {
+            response,
+            state_update,
+            _state: core::marker::PhantomData,
+        } = self;
+
+        state_update(state).await;
+
+        response
+            .write_to_with_state(state, connection, response_writer)
+            .await
+    }
+}
+
+/// An extension trait for updating the state as part of writing the response.
+///
+/// Allows for easy state updates using data produced by the handler.
+pub trait WithStateUpdate<State>: IntoResponseWithState<State> {
+    fn with_state_update<F: AsyncFnOnce(&State)>(
+        self,
+        state_update: F,
+    ) -> IntoResponseWithStateUpdate<State, Self, F> {
+        IntoResponseWithStateUpdate {
+            response: self,
+            state_update,
+            _state: core::marker::PhantomData,
+        }
+    }
+}
+
+impl<State, T: IntoResponseWithState<State>> WithStateUpdate<State> for T {}